+++ CHANGELOG.rst
@@ -4,16 +4,18 @@ Changelog for package rodney
 
 Unreleased
 ------------------
+* Added IMU calibration node to launch file
 * Added configuration and launching of the nav stack
 * Added remote heartbeat message which will be run on remote workstation when teleoping
 * Added RPLidar and scan filter nodes to rodney launch file
 * Added functionality to manually enable the LIDAR from keyboard or joystick
 * Added LIDAR to URDF model
 * Added rodney.xacro and now auto generate rodney.urdf file
-* "Arduino" code changed to use either an Arduino Nano or a Teensy. Nano uses smaller buffer sizes a Teensy use higher baud rate
-* Increased tacho message publish rate to 40Hz on a Teensy
+* "Arduino" code changed to use a Teensy and now reads IMU data
 * Changed head joints in URDF model from EfforJointInterface to PositionJointInterface
 * Moved scripts to a scripts folder and create rules for remapping serial ports
+* Now using use_ramp parameter instead of use_pid
+
 
 0.3.2 (2019-03-19)
 ------------------